split source file
